[DF-6639] Fixed x64 warning
Former-commit-id: 5fc032d3a380f1dec173509a34d6f50b892926ce

+++ Sources/Internal/Scene3D/Components/VisibilityCheckComponent.cpp
@@ -146,7 +146,7 @@ void VisibilityCheckComponent::BuildPointSet()
         while (canGenerate)
         {
             canGenerate = false;
-            for (int i = points.size() - 1; i >= 0; --i)
+            for (int i = static_cast<int>(points.size()) - 1; i >= 0; --i)
             {
                 if (generateAroundPoint(points.at(i)))
                 {
